% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMHalf.R
\name{ADMMHalf}
\alias{ADMMHalf}
\title{ADMMHalf - alternating direction method of multipliers with half thresholding operator}
\usage{
ADMMHalf(A, b, x0, sparsity, MaxIter = 200, epsilon = 1e-09)
}
\arguments{
\item{A}{Gene expression data of transcriptome factors (i.e. basis function in machine learning).}

\item{b}{Gene expression data of target genes (i.e. observation in machine learning).}

\item{x0}{Gene expression data of Chromatin immunoprecipitation or zero vector (i.e. initial iterative point in machine learning).}

\item{sparsity}{Sparsity level of solution.}

\item{MaxIter}{Maximum iteration used in calculation; we set default as 200.}

\item{epsilon}{Stopping rule in algorithm where the rule is \eqn{||Ax-b|{|_2^2} < \epsilon}; we set default as 1e-9.}
}
\description{
This function is the process of ADMM-Half algorithm aims to solve \eqn{l_{\frac{1}{2}}} regularization optimization model.
}
\details{
ADMM-Half is an extension of ADMM, where the half thresholding operator is adopted to replace the soft one in ADMM. It is aimed to solve the \eqn{l_{\frac{1}{2}}} regularization problem:
\deqn{{\min}_{x \in R^n} ||Ax - b|{|_2^2} + \lambda ||x|{|_{1/2}^{1/2}}}
}
\author{
Yaohua Hu <mayhhu@szu.edu.cn>

Xinlin Hu <ttxinlinhu@qq.com>
}
