% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMHard.R
\name{ADMMHard}
\alias{ADMMHard}
\title{ADMMHard - alternating direction method of multipliers with hard thresholding operator}
\usage{
ADMMHard(A, b, x0, sparsity, MaxIter = 200, epsilon = 1e-09)
}
\arguments{
\item{A}{Gene expression data of transcriptome factors (i.e. basis function in machine learning).}

\item{b}{Gene expression data of target genes (i.e. observation in machine learning).}

\item{x0}{Gene expression data of Chromatin immunoprecipitation or zero vector (i.e. initial iterative point in machine learning).}

\item{sparsity}{Sparsity level of solution.}

\item{MaxIter}{Maximum iteration used in calculation; we set default as 200.}

\item{epsilon}{Stopping rule in algorithm where the rule is \eqn{||Ax-b|{|_2^2} < \epsilon}; we set default as 1e-9.}
}
\description{
This function is the process of ADMM-Hard algorithm aims to solve \eqn{l_0} regularization optimization model.
}
\details{
ADMM-Hard was introduced by Boyd et al.(2011). It is an extension of ADMM, where the hard thresholding operator is adopted to replace the soft one in ADMM. It is aimed to solve the \eqn{l_0} regularization problem:
\deqn{{\min}_{x \in R^n} ||Ax - b|{|_2^2} + \lambda ||x|{|_0}}
}
\author{
Yaohua Hu <mayhhu@szu.edu.cn>

Xinlin Hu <ttxinlinhu@qq.com>
}
