% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(datexpr, num = 5, seed = TRUE)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Numeric matrix with the dataset to infer the network.Rows contain genes and columns contain samples}

\item{num}{Number of the times of resampling with bootstrap method.The number of new datasets.Default: 5.}

\item{seed}{If set to TRUE, set.seed for keeping the results the same for each run. Default: TRUE.}
}
\value{
A list for the bootstrap result of gene expression matrix. Each one of objects contains a matrix of the resampled dataset.
}
\description{
The gene expression matrix was resampled by bootstrap method.
}
\details{
Bootstrap is a resampling technique in statistical learning, which is used to estimate standard error, confidence interval and deviation. We resampled the gene expression matrix using bootstrap method to form multiple datasets.
}
\examples{
data(datalist)
datexpr <- datalist$B[1:10,1:10]
boot.sample <- bootstrap(datexpr)
}
