% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calauc.R
\name{calauc}
\alias{calauc}
\title{Calculate the value of AUROC and AUPR.}
\usage{
calauc(dat)
}
\arguments{
\item{dat}{A data.frame. The first element, dat$predictions, is a vector of numerical predictions. The second element, dat$labels, is a vector of cordatponding class labels.}
}
\value{
A vector of the value of AUROC and AUPR.
}
\description{
Calculate the value of the area under the ROC-curve and the value of the area under the PR-curve.
}
\details{
AUROC:

Area under the ROC curve. This is equal to the value of the Wilcoxon-Mann-Whitney test statistic and also the probability that the classifier will score are randomly drawn positive sample higher than a randomly drawn negative sample.

AUPR:

Area under the Precision/Recall curve.
}
\examples{
dat <- data.frame(predictions = runif(200, min=0, max=1),labels = c(rep(1,100),rep(0,100)))
auc <- calauc(dat)
print(auc)
}
\references{
Fawcett, T. 2004. "ROC Graphs: Notes and Practical Considerations for Researchers." In HPL-2003-4, 89-96. HP Labs, Palo Alto, CA.

Sing, Tobias, Niko Beerenwinkel, and Thomas Lengauer. 2004.
 "Learning Mixtures of Localized Rules by Maximizing the Area Under the Roc Curve."
 In In et Al Jose Hernandez-Orallo, Editor, 1st International Workshop on Roc Analysis in Artificial Intelligence, 89-96.
}
