% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calpl.R
\name{calpl}
\alias{calpl}
\title{Calculate the predictions of GRN and the labels of the high-throughput golden standard.}
\usage{
calpl(HGSmat, grn, sym = TRUE, num = NULL)
}
\arguments{
\item{HGSmat}{A matrix which is the high-throughput golden standard. Rows contain genes and columns contain TFs ,the value between in one or zero.}

\item{grn}{A matrix which is the weighted adjacency matrix of the inferred GRN by this algorithm or the result from the function of out_grn. The matrix need the colnames(TF) and rownames(genes).}

\item{sym}{If set to TRUE, only the regulatory relationship in the high-throughput gold standard is considered, and the value of 0 is not considered, so that the number of true positive and false positive is equal.
If set to FALSE, It is assumed that the regulatory relationships that do not exist in the high-throughput gold standard are non regulatory and 0.Default: TRUE.}

\item{num}{The number of gene regulator information. Default: NULL}
}
\value{
A data.frame. The first element, dat$predictions, is a vector of numerical predictions. The second element, dat$labels, is a vector of class labels.
}
\description{
Using the gold standard data to evaluate the predicted GRN. Calculate the predictions of GRN and the labels of the high-throughput golden standard.
}
\examples{
A <- matrix(rnorm(200,0,1),20,10)
B <- matrix(rnorm(1000,0,1),100)
X <- matrix(0,100,20)
s <- c(1:10)
TF <- paste0('G',1:20)
gene <- paste0('G',1:100)
res <- demo_reg(A, B, s, X, 'ADMM',max.steps = 200, TF, gene, cl.core = 2,file=NULL,verbose=FALSE)
grn <- res$grn
HGSmat <- matrix(0,100,20)
rownames(HGSmat) <- gene
colnames(HGSmat) <- TF
HGSmat[sample(2000,200)] <- 1
dat <- calpl(HGSmat,grn,sym=FALSE)
head(dat)
}
