% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{Confusion Matrix}
\usage{
confusion(dat, threshold = 0)
}
\arguments{
\item{dat}{A data.frame. The first element, dat$predictions, is a vector of numerical predictions. The second element, dat$labels, is a vector of cordatponding class labels.}

\item{threshold}{Between 0 and 1. If their weight value is over the threshold, they are considered to have a regulatory relationship, otherwise they are not considered to have regulatory relationship. Default: 0.}
}
\value{
A vector of confusion matrix including some evalution parameters. (TP, FP, FN, precision, recall, F1-score).
}
\description{
Calculate the confusion martix, takes the weighted adjacency matrix of the inferred network by this algorithm and the true or high-throughput golden standard network for validation.
}
\examples{
dat <- data.frame(predictions = runif(200, min=0, max=1),labels = c(rep(1,100),rep(0,100)))
score <- confusion(dat)
print(score)

}
\references{
G. Altay, F. Emmert-Streib, "Inferring the conservative causal core of gene regulatory networks", BMC Systems Biology, (2010) 4:132.
}
