% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_kendall.R
\name{cor_kendall}
\alias{cor_kendall}
\title{Kendall correlation}
\usage{
cor_kendall(datexpr, gene, TF = NULL, file = NULL)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}
}
\value{
A matrix which is the weighted adjacency matrix of the inferred network by this algorithm.
}
\description{
Default parameters for kendall algorithm.
}
\details{
The weighted adjacency matrix of the inferred network by this algorithm keep the value.
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:10]
gene <- datalist$gene[1:100]
grn <- cor_kendall(datexpr,gene)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988). The New S Language. Wadsworth & Brooks/Cole.

Kendall, M. G. (1938). A new measure of rank correlation, Biometrika.

Kendall, M. G. (1945). The treatment of ties in rank problems. Biometrika.
}
