% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_pearson.R
\name{cor_pearson}
\alias{cor_pearson}
\title{Pearson correlation}
\usage{
cor_pearson(datexpr, gene, TF = NULL, file = NULL)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}
}
\value{
A matrix which is the weighted adjacency matrix of the inferred network by this algorithm.
}
\description{
Default parameters for pearson algorithm.
}
\details{
The weighted adjacency matrix of the inferred network by this algorithm keep the value which the p-value<0.05.
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- cor_pearson(datexpr,gene)
}
\references{
Hollander M. and Wolfe D.A. (1973). Nonparametric Statistical Methods. New York: Wiley.

Press WH, Flannery BP, Teukolsky SA, Vetterling, WT (1988): Numerical Recipes in C. Cambridge: Cambridge University Press.
}
