% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSaMP.R
\name{CoSaMP}
\alias{CoSaMP}
\title{CoSaMP - compressive sampling matched pursuit}
\usage{
CoSaMP(A, b, sparsity, MaxIter = 200, epsilon1 = 1e-06, epsilon2 = 1e-06)
}
\arguments{
\item{A}{Gene expression data of transcriptome factors (i.e. basis function in machine learning).}

\item{b}{Gene expression data of target genes (i.e. observation in machine learning).}

\item{sparsity}{Sparsity level of solution.}

\item{MaxIter}{Maximum iteration used in calculation; we set default as 200.}

\item{epsilon1}{First stopping rule in algorithm where the rule is \eqn{Ax-b < \epsilon_1}; we set default as 1e-6.}

\item{epsilon2}{Second stopping rule in algorithm where the rule is \eqn{\max{|A'(Ax-b)|} < \epsilon_2}; we set default as 1e-6.}
}
\description{
This function is the process of CoSaMP (Compressive sampling matched pursuit).
}
\details{
Compressive sampling matched pursuit (CoSaMP) algorithm is an approximation algorithm proposed by Needell and Tropp (2009) for solving problem:
\deqn{\begin{array}{l}\min ||Ax-b|{|^2}\\s.t.||x|{|_0} \le s\end{array}}

CoSaMP is ultimately based on OMP, while it incorporates the sampling technique to accelerate the algorithm and to enjoy strong theoretical guarantees than OMP. At each iteration of CoSaMP, some random samples are added into the set of selected features, and then a least squares scheme is employed to estimate the approximation from the set of selected features. This process is repeated until the recoverable features have been found. CoSaMP shares the fast implementation advantage and remedies the correction inability of OMP.
}
\references{
Needell, D. and Tropp, J. (2009). "CoSaMP: Iterative signal recovery from incomplete and inaccurate samples", Applied and Computational Harmonic Analysis, 26(3), 301-321.
}
\author{
Yaohua Hu <mayhhu@szu.edu.cn> 

Xinlin Hu <ttxinlinhu@qq.com>
}
