% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_other.R
\name{demo_other}
\alias{demo_other}
\title{Demo other - The demo of other method}
\usage{
demo_other(
  datexpr,
  algorithm = "Pearson",
  TF = NULL,
  gene,
  cl.cores = 2,
  file = NULL,
  verbose = TRUE
)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{algorithm}{The algorithm of inferring grn. We provide 19 algorithms in this function including 'Pearson','Spearman','Kendall','WGCNA','BC3NET','C3NET',
'parmigene_ARACNE.a','parmigene_ARACNE.m','parmigene_CLR','parmigene_MRNET','minet_ARACNE','minet_CLR','minet_MRNET','minet_MRNETB','GeneNet',
'GENIE3.ET','GENIE3.RF','ENNET,'PCIT'. Default: Pearson.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{gene}{A vector of characters containing the target genes.}

\item{cl.cores}{The number of cores in computer that you want to utilize. Default: 2.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}

\item{verbose}{If set to TRUE, a feedback on the progress of the progress of the calculations is given. Default: TRUE.}
}
\value{
A list with the function of parameters ,algorithms, elapse time, grn which a matrix which is the weighted adjacency matrix of the network inferred by the this algorithm
}
\description{
This is the main function to call different algorithms for network inference.
}
\details{
The demo function aims to call algorithms to infer gene regulatory network.
}
\examples{
datexpr <- matrix(rnorm(1000,0,1),100)
TF <- paste0('G',1:20)
gene <- paste0('G',1:100)
res <- demo_other(datexpr=datexpr,algorithm = 'Pearson' ,TF=TF,gene=gene,file=NULL,verbose=TRUE)
grn <- res$grn
grn[1:10,1:10]
}
