% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_plot.R
\name{draw_plot}
\alias{draw_plot}
\title{draw_plot}
\usage{
draw_plot(dat, method = "ROC", algname)
}
\arguments{
\item{dat}{A data.frame. The first element, dat$predictions, is a vector of numerical predictions. The second element, dat$labels, is a vector of cordatponding class labels.}

\item{method}{Receiveroperating characteristic curve, precision-recall curve and sensitivity-specificity curve (ROC,PR, SSC). Default: ROC.}

\item{algname}{The algorithm of gene network inferring.}
}
\value{
The curve which selected, such as ROC and PR.
}
\description{
Draw the curve for the gene regulatory network.
}
\examples{
set.seed(1)
dat <- data.frame(predictions = runif(200, min=0, max=1),labels = c(rep(1,100),rep(0,100)))
p <- draw_plot(dat = dat,method = 'ROC',algname ='ADMM')
print(p)
}
