% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGSt.R
\name{HGSt}
\alias{HGSt}
\title{Transform the high-throughput gold standard data}
\usage{
HGSt(TF, gene, HGS)
}
\arguments{
\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{gene}{A vector of characters containing the target genes.}

\item{HGS}{The class of HGS is 'data.frame'. The golden standard gene regulator network. The first column is TFs list and the second one is target genes list. 
The value of third column should be 1 or 0.}
}
\value{
A matrix which is the high-throughput golden standard. Rows contain genes and columns contain TFs ,the value between in one or zero.
}
\description{
The high-throughput gold standard data is represented in matrix
}
\details{
The three columns of high-throughput gold standard data are transformed into matrix to evaluate the inferred gene regulatory network.
}
\examples{
data(datalist)
HGSmat <- HGSt(datalist$TF,datalist$gene,datalist$HGS)
}
