% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISTA.R
\name{ISTA}
\alias{ISTA}
\title{ISTA - iterative soft thresholding algorithm}
\usage{
ISTA(A, b, x0, sparsity, MaxIter = 200, epsilon = 1e-06)
}
\arguments{
\item{A}{Gene expression data of transcriptome factors (i.e. basis function in machine learning).}

\item{b}{Gene expression data of target genes (i.e. observation in machine learning).}

\item{x0}{Gene expression data of Chromatin immunoprecipitation or zero vector (i.e. initial iterative point in machine learning).}

\item{sparsity}{Sparsity level of solution.}

\item{MaxIter}{Maximum iteration used in calculation; we set default as 200.}

\item{epsilon}{Stopping rule in algorithm where the rule is \eqn{||Ax-b|{|_2^2} < \epsilon}; we set default as 1e-6.}
}
\description{
This function is the process of ISTA algorithm aims to solve \eqn{l_1} regularization optimization model.
}
\details{
Iterative soft thresholding algorithm (ISTA) was introduced by Daubechies et al. (2004) for solving the \eqn{l_1} regularization problem:
\deqn{{\min}_{x \in R^n} ||Ax - b|{|_2^2} + \lambda ||x|{|_1}}
The idea of ISTA is to sequentially proceed a gradient descent step and a soft thresholding operator at each iteration.
}
\references{
Daubechies, I., Defrise, M., and De Mol, C. (2004). "An iterative thresholding algorithm for linear inverse problems with a sparsity constraint", Communications on Pure and Applied Mathematics, 57, 1413-1457.
}
\author{
Yaohua Hu <mayhhu@szu.edu.cn> 

Xinlin Hu <ttxinlinhu@qq.com>
}
