% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minet_mrnet.R
\name{minet_mrnet}
\alias{minet_mrnet}
\title{Maximum Relevance Minimum Redundancy algorithm (from R package minet)}
\usage{
minet_mrnet(datexpr, gene, TF = NULL, file = NULL)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}
}
\value{
A matrix which is the weighted adjacency matrix of the network inferred by aracne algorithm. Rows contain genes and columns contain TF.
}
\description{
Default parameters for MRNET network inference algorithm.
}
\details{
MRNET algorithm Gene regulatory network inference
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- minet_mrnet(datexpr,gene)
}
\references{
H. Peng, F.long and C.Ding. Feature selection based on mutual information: Criteria of max-dependency, max relevance and min redundancy. IEEE transaction on Pattern Analysis and Machine Intelligence, 2005.

Patrick E. Meyer, Frederic Lafitte and Gianluca Bontempi. minet: A R/Bioconductor Package for Inferring Large Transcriptional Networks Using Mutual Information. BMC Bioinformatics, Vol 9, 2008.
}
