% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minet_mrnetb.R
\name{minet_mrnetb}
\alias{minet_mrnetb}
\title{Maximum Relevance Minimum Redundancy Backward (from R package minet)}
\usage{
minet_mrnetb(datexpr, gene, TF = NULL, file = NULL)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}
}
\value{
A matrix which is the weighted adjacency matrix of the network inferred by MRNETB algorithm.
}
\description{
Default parameters for MRNETB network inference algorithm.
}
\details{
MRNETB takes the mutual information matrix as input in order to infer the network using the maximum relevance/minimum redundancy criterion combined with a backward elimination and a sequential replacement - see references. This method is a variant of mrnet.
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- minet_mrnetb(datexpr,gene)
}
\references{
Patrick E. Meyer, Daniel Marbach, Sushmita Roy and Manolis Kellis. Information-Theoretic Inference of Gene Networks Using Backward Elimination. The 2010 International Conference on Bioinformatics and Computational Biology.

Patrick E. Meyer, Kevin Kontos, Frederic Lafitte and Gianluca Bontempi. Information-theoretic inference of large transcriptional regulatory networks. EURASIP Journal on Bioinformatics and Systems Biology, 2007.
}
