% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMP.R
\name{OMP}
\alias{OMP}
\title{OMP - orthogonal matching pursuit}
\usage{
OMP(A, b, sparsity, epsilon1 = 1e-06, epsilon2 = 1e-06)
}
\arguments{
\item{A}{Gene expression data of transcriptome factors (i.e. basis function in machine learning).}

\item{b}{Gene expression data of target genes (i.e. observation in machine learning).}

\item{sparsity}{Sparsity level of solution.}

\item{epsilon1}{First stopping rule in algorithm where the rule is \eqn{Ax-b < \epsilon_1}; we set default as 1e-6.}

\item{epsilon2}{Second stopping rule in algorithm where the rule is \eqn{\max{|A'(Ax-b)|} < \epsilon_2}; we set default as 1e-6.}
}
\description{
This function is the process of OMP (orthogonal matching pursuit).
}
\details{
Orthogonal matching pursuit (OMP) algorithm was introduced by Tropp (2004) for the recovery of a high-dimensional sparse signal based on a small number of noisy linear measurements.
OMP is a kind of forward greedy methods for approaching an approximate solution of problem:
\deqn{\begin{array}{l}\min ||Ax - b|{|^2}\\s.t.||x|{|_0} \le s\end{array}}
OMP is a kind of forward greedy methods for approaching an approximate solution of convex problems. The idea of OMP is to iteratively select a column of A that is most correlated with the current residuals, and hence to make most significant progress at each iteration in order to reduce the loss function and achieve sparsity.
}
\references{
Tropp, J. A. (2004). "Greed is good: algorithmic results for sparse approximation", IEEE Transactions on Information Theory, 50(10), 2231-2242.
}
\author{
Yaohua Hu <mayhhu@szu.edu.cn> 

Xinlin Hu <ttxinlinhu@qq.com>
}
