% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_grn.R
\name{out_grn}
\alias{out_grn}
\title{Output grn}
\usage{
out_grn(grn, num = NULL, TF, gene, file = NULL)
}
\arguments{
\item{grn}{A matrix which is the weighted adjacency matrix of the inferred network by this algorithm.}

\item{num}{The number of gene regulator imformation. Default: NULL.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{gene}{A vector of characters containing the target genes.}

\item{file}{A connection, or a character string naming the file to write to. The default not save. Default: NULL.}
}
\value{
A data.frame. A ranked edge list. The first column is TF, the second is gene, and the last columns is the weighted adjacency values.
}
\description{
Converted to a ranked edge list
}
\details{
The output is consistent with the data format required by Cytoscape. This output data can be used to map the gene regulatory network with Cytoscape.
}
\examples{
B <- matrix(rnorm(1000,0,1),100)
TF <- paste0('G',1:20)
gene <- paste0('G',1:100)
grn <- run_aracne.a(B,gene,TF)
grn <- out_grn(grn,TF=TF,gene=gene)
head(grn)
}
