% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aracne.a.R
\name{run_aracne.a}
\alias{run_aracne.a}
\title{ARACNE algorithm (additive model)}
\usage{
run_aracne.a(datexpr, gene, TF = NULL, file = NULL)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}
}
\value{
A matrix which is the weighted adjacency matrix of the network inferred by aracne algorithm. Rows contain genes and columns contain TF.
}
\description{
Default parameters for ARACNE algorithm for the reconstruction of gene interaction networks (additive model).
}
\details{
The motivation of the Algorithm for the Reconstruction of Accurate Cellular Networks (ARACNE) is that many similar measures between variables may be the result of indirect effects. In order to delete the indirect effect the algorithm relies on the "Data Processing Inequality", this process removes the weakest link in every triplet of connected variables.
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- run_aracne.a(datexpr,gene)
}
\references{
Adam A. Margolin, Ilya Nemenman, Katia Basso, Chris Wiggins, Gustavo Stolovitzky, Riccardo Dalla Favera, and Andrea Califano. Aracne : An algorithm for the reconstruction of gene regulatory networks in a mammalian cellular context. BMC Bioinformatics, 2006.
}
