% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_clr.R
\name{run_clr}
\alias{run_clr}
\title{Context Likelihood or Relatedness Network}
\usage{
run_clr(datexpr, gene, TF = NULL, file = NULL)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}
}
\value{
Matrix returns a matrix which is the weighted adjacency matrix of the network inferred by CLR algorithm.
}
\description{
Default parameters for CLR network inference algorithm.
}
\details{
CLR Gene regulatory network inference.
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- run_clr(datexpr,gene)
}
\references{
Jeremiah J. Faith, Boris Hayete, Joshua T. Thaden, Ilaria Mogno, Jamey Wierzbowski, Guillaume Cottarel, Simon Kasif, James J. Collins, and Timothy S. Gardner. Large-scale mapping and validation of escherichia coli transcriptional regulation from a compendium of expression profiles. PLoS Biology, 2007.
}
