% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ennet.R
\name{run_ennet}
\alias{run_ennet}
\title{Gene regulatory network inference for algorithm of ENNET}
\usage{
run_ennet(datexpr, gene, TF = NULL, file = NULL, cl.cores = 2)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}

\item{cl.cores}{The number of cores in computer that you want to utilize. Default: 2.}
}
\value{
A matrix which is the weighted adjacency matrix of the network inferred by ENNET algorithm.
}
\description{
Default parameters for ENNET network inference algorithm.
}
\details{
ENNET Gene regulatory network inference
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- run_ennet(datexpr,gene)
}
\references{
Slawek, J., Arodz, T. ENNET: inferring large gene regulatory networks from expression data using gradient boosting. BMC Syst Biol 7, 106 (2013). https://doi.org/10.1186/1752-0509-7-106
}
