% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Genie3.RF.R
\name{run_GENIE3.RF}
\alias{run_GENIE3.RF}
\title{GENIE3.RF}
\usage{
run_GENIE3.RF(datexpr, gene, TF = NULL, file = NULL, cl.cores = 2)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}

\item{cl.cores}{The number of cores in computer that you want to utilize. Default: 2.}
}
\value{
Matrix returns a matrix which is the weighted adjacency matrix of the network inferred by GENIE3 algorithm (Random Forests).
}
\description{
Default parameters for GENIE3 algorithm (Random Forests)
}
\details{
GENIE3 Infers a gene regulatory network (in the form of a weighted adjacency matrix) from expression data, using ensembles of regression trees(Random Forests).
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
TF <- gene[1:20]
#grn <- run_GENIE3.RF(datexpr,gene,TF)
}
\references{
Irrthum, Alexandre, Louis Wehenkel, and Pierre Geurts. "Inferring regulatory networks from expression data using tree-based methods." PloS one 5.9 (2010): e12776.

Breiman, Leo. "Random forests." Machine learning 45.1 (2001): 5-32.
}
