% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pcit.R
\name{run_pcit}
\alias{run_pcit}
\title{PCIT algorithm}
\usage{
run_pcit(datexpr, gene, TF = NULL, file = NULL)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}
}
\value{
A matrix which is the weighted adjacency matrix of the network inferred by PCIT algorithm.
}
\description{
Default parameters for PCIT algorithm
}
\details{
The Partial Correlation coefficient with Information Theory (PCIT) algorithm, combines the concept of partial correlation coefficient with information theory to identify significant gene-to-gene associations.
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- run_pcit(datexpr,gene)
}
\references{
Reverter, Antonio, and Eva KF Chan. "Combining partial correlation and an information theory approach to the reversed engineering of gene co-expression networks." Bioinformatics 24.21 (2008): 2491-2497.
}
