% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_WGCNA.R
\name{run_wgcna}
\alias{run_wgcna}
\title{WGCNA Weighted Correlation Network}
\usage{
run_wgcna(datexpr, gene, TF = NULL, file = NULL, cl.cores = 2)
}
\arguments{
\item{datexpr}{Expression matrix (genes * samples). Every row is a gene, every column is a sample.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}

\item{cl.cores}{The number of cores in computer that you want to utilize. Default: 2.}
}
\value{
A matrix which is the correlation matrix by WGCNA algorithm.
}
\description{
Default parameters for WGCNA algorithm
}
\details{
WGCNA Gene regulatory network inference
}
\examples{
data(datalist)
datexpr <- datalist$B[1:100,1:100]
gene <- datalist$gene[1:100]
grn <- run_wgcna(datexpr,gene)
}
\references{
Bin Zhang and Steve Horvath (2005) "A General Framework for Weighted Gene Co-Expression Network Analysis", Statistical Applications in Genetics and Molecular Biology: Vol. 4: No. 1, Article 17

Nowick K, Gernat T, Almaas E, Stubbs L. Differences in human and chimpanzee gene expression patterns define an evolving network of transcription factors in brain. Proc Natl Acad Sci U S A. 2009 Dec 29;106(52):22358-63. doi: 10.1073/pnas.0911376106. Epub 2009 Dec 10.

Gysi DM, Voigt A, Fragoso TM, Almaas E, Nowick K. wTO: an R package Wfor computing weighted topological overlap and a consensus network with integrated visualization tool. BMC Bioinformatics. 2018 Oct 24;19(1):392. doi: 10.1186/s12859-018-2351-7.
}
