% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGL0.R
\name{SPGL0}
\alias{SPGL0}
\title{SPGL0 - spectral projected gradient method to solve L0 regularization problem}
\usage{
SPGL0(A, b, x0, sparsity, MaxIter = 200, epsilon = 1e-06)
}
\arguments{
\item{A}{Gene expression data of transcriptome factors (i.e. basis function in machine learning).}

\item{b}{Gene expression data of target genes (i.e. observation in machine learning).}

\item{x0}{Gene expression data of Chromatin immunoprecipitation or zero vector (i.e. initial iterative point in machine learning).}

\item{sparsity}{Sparsity level of solution.}

\item{MaxIter}{Maximum iteration used in calculation; we set default as 200.}

\item{epsilon}{Stopping rule in algorithm where the rule is \eqn{||Ax-b|{|_2^2} < \epsilon}; we set default as 1e-6.}
}
\description{
This function is the process of SPGL0 algorithm aims to solve \eqn{l_0} regularization optimization model, and it can be understood as an extension of SPGL1.
}
\details{
Projected gradient method (named as M-sparse algorithm) was introduced by Blumensath and Davies (2008). Inspired by the idea of Barzilai and Borwein stepsize rule Barzilai and Borwein (1988), we here propose and apply the SGPL0 (spectral projected gradient method) to solve the \eqn{l_0} regularization problem:
\deqn{{\min}_{x \in R^n} ||Ax - b|{|_2^2} + \lambda ||x|{|_0}}
The only difference from the M-sparse algorithm is to use the Barzilai-Borwein stepsize rule replace the constant stepsize. SPGL0 can also be understood as an extension of SGPL1, in which we use projection onto M coefficients with the largest magnitude instead of the projection onto the \eqn{l_1} norm ball.
}
\references{
Blumensath, T. and Davies, M. E. (2008). "Iterative thresholding for sparse approximations", Journal of Fourier Analysis and Applications, 14, 629-654.

Barzilai, J. and Borwein, J. M. (1988). "Two-point step size gradient methods", IMA Journal of Numerical Analysis, 8(1),141-148.
}
\author{
Yaohua Hu <mayhhu@szu.edu.cn> 

Xinlin Hu <ttxinlinhu@qq.com>
}
